的服务层实现

+++ src/main/java/com/example/auctionplatform/dao/AuctionItemRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuctionItemRepository extends CrudRepository<AuctionItem, Long> {
+public interface AuctionItemRepository extends CrudRepository<AuctionItem, Integer> {
 }
